Create stives.py
This is my first attempt to code the St. Ives Riddle to achieve a correct answer if we flatten out all of the assumptions to true. 

Mind you the popular answer to this riddle is "1"

I'd argue that the correct answer is "Unknown" and that even "0" is closer to correct. For the same reason I include sacks, I do not include the narrator, because he/she isn't asked for in the question "kits, cats, sacks, and wives" even if we assume the narrator is still going.

If anyone wants to help make this better, I applaud you for having more free time than I do. I would love to see this turned into an RPG where we can have functions and prompts so that there can be a hosted app where you can ask people to contemplate the variables and try it on many different ways.

I've clearly gone too far down this rabbit hole, since it's Monday and this was my Saturday distraction exercise. I'm going to post and leave it up to the world to do with this what you want. Feel free to e-mail me if you have any questions.